% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creationMatrice.R
\name{creationMatrice}
\alias{creationMatrice}
\title{creationMatrice}
\usage{
creationMatrice()
}
\value{
a 9 by 9 matrix (a full sudoku grid)
}
\description{
this function create a 9 by 9  matrix in wich numbers appears once by columns, by row and by 3x3 boxes
}
\details{
Create a full sudoku grid
}
\examples{
creationMatrice()
}
\author{
Chloe Mason
}
