% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creationNiveau.R
\name{creationNiveau}
\alias{creationNiveau}
\title{creationNiveau}
\usage{
creationNiveau(m, difficulte)
}
\arguments{
\item{m}{a 9 by 9 matrix (sudoku grid)}

\item{difficulte}{a string refering to the difficulty, the values should be : "facile", "moyen" or "difficile"}
}
\value{
The sudoku grid with empty boxes (number equals to 0)
}
\description{
Add together two numbers
}
\examples{
m<-creationMatrice()
creationNiveau(m, "facile")
}
\author{
Chloe Mason
}
