% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trouveSolution.R
\name{trouveSolution}
\alias{trouveSolution}
\title{trouveSolution}
\usage{
trouveSolution(m)
}
\arguments{
\item{m}{a 9 by 9 matrix}
}
\value{
a string explaining if a solution exists
}
\description{
Given a matrix, find if a solution exists
}
\details{
A function to find if a solution for a sudoku grid
}
\examples{
m<-creationMatrice()
trouveSolution(m)
}
